fixed issue where normalize_matrix was not defin ed in Optimization.py. Added a helper.py file and moved functions that werent calling the clas ses self.{data} I only did this for optimizatoin.py because I saw functions that were listed a s undefined and werent calling the classes values but rather just returning data and they coul d be classified as general functions that other classes could use to avoid repeated code or to  better strucute the file this could maybe be done withing other class definitions as well. Ad ded testAdamTraining in tests.py to do a minimal test on the full inversion optimization metho d. Removed two instances of the same function in Optimization.py
