+++ OpenMiChroM/Optimization.py
@@ -27,6 +27,7 @@ from sklearn.preprocessing import normalize
 import os
 import pandas as pd
 import warnings
+import helper
 
 class AdamTraining:
     R"""
@@ -93,46 +94,6 @@ class AdamTraining:
         self.t += 1
         return w, b
 
-    def knight_ruiz_balance(matrix, tol=1e-5, max_iter=100):
-        R"""
-        Perform the Knight-Ruiz matrix balancing.
-        """
-        A = np.array(matrix, dtype=float)
-        n = A.shape[0]
-        row_scaling = np.ones(n)
-        col_scaling = np.ones(n)
-
-        for _ in range(max_iter):
-            row_scaling = np.sqrt(np.sum(A, axis=1))
-            A /= row_scaling[:, None]
-            col_scaling = np.sqrt(np.sum(A, axis=0))
-            A /= col_scaling
-
-            if np.all(np.abs(row_scaling - 1) < tol) and np.all(np.abs(col_scaling - 1) < tol):
-                break
-
-        return A
-
-
-    def normalize_matrix(matrix):
-        R"""
-        Normalize the matrix for simulation optimization. Here the first neighbor should have the probability of contact P=1.0.
-        """
-        matrix = np.nan_to_num(matrix, nan=0, posinf=0, neginf=0)
-        np.fill_diagonal(matrix,0.0)
-
-        max_values = np.amax(np.triu(matrix), axis=1)
-        
-        # To avoid division by zero, replace zeros with ones
-        max_values[max_values == 0] = 0.0000001
-        
-        normalized_matrix = np.triu(matrix) / max_values[:, np.newaxis]
-        # return normalized_matrix
-        matrix= normalized_matrix + np.triu(normalized_matrix,k=1).T
-        np.fill_diagonal(matrix,1.0)
-
-        return matrix
-
 
     def getPars(self, HiC, centerRemove=False, centrange=[0,0], cutoff='deprecate', norm=True, cutoff_low=0.0, cutoff_high=1.0, KR=False, neighbors=0):
         R"""
@@ -159,10 +120,10 @@ class AdamTraining:
             allmap = np.loadtxt(HiC)
 
         if KR==True:
-            allmap = knight_ruiz_balance(allmap)
+            allmap = helper.knight_ruiz_balance(allmap)
 
         if norm==True:
-            r=normalize_matrix(allmap)
+            r=helper.normalize_matrix(allmap)
 
             for i in range(len(r)-1):
                 maxElem = r[i][i+1]
@@ -243,10 +204,10 @@ class AdamTraining:
             allmap = np.loadtxt(HiC)
 
         if KR==True:
-            allmap = knight_ruiz_balance(allmap)
+            allmap = helper.knight_ruiz_balance(allmap)
 
         if norm==True:
-            r=normalize_matrix(allmap)
+            r=helper.normalize_matrix(allmap)
 
             for i in range(len(r)-1):
                 maxElem = r[i][i+1]
@@ -407,7 +368,7 @@ class FullTraining:
             lambdas[i] = initialGuess
         lambdas = np.triu(lambdas) + np.triu(lambdas).T
 
-        self.saveLambdas(sequenceFile=sequenceFile, data=lambdas, outputPath=outputPath, name="lambda_0")
+        helper.saveLambdas(sequenceFile=sequenceFile, data=lambdas, outputPath=outputPath, name="lambda_0")
 
     def appCutoff(self, pair_h, c_h, pair_l, c_l):
         R"""
@@ -505,12 +466,7 @@ class FullTraining:
 
 
         return(pearsonr(a1,a2)[0])
-    def saveLambdas(self, sequenceFile, data, outputPath, name):
-        seq = np.loadtxt(sequenceFile, dtype=str)[:,1]
 
-        lamb = pd.DataFrame(data,columns=seq)
-        lamb.to_csv(os.path.join(outputPath, name), index=False)
-        print("{} file save in {}".format(name, outputPath))
 
         
     def getLambdas(self):
@@ -570,7 +526,7 @@ class CustomMiChroMTraining:
     """
    
     def __init__(self, ChromSeq="chr_beads.txt", TypesTable=None, mu=3.22, rc=1.78, cutoff=0.0, IClist=None, dinit=3, dend=200): 
-        self.ChromSeq = self._get_chrom_seq(ChromSeq)
+        self.ChromSeq = helper.get_chrom_seq(ChromSeq)
         self.size = len(self.ChromSeq)
         self.P=np.zeros((self.size,self.size))
         self.Pold=np.zeros((self.size,self.size))
@@ -608,29 +564,7 @@ class CustomMiChroMTraining:
             except IOError:
                 print("Error in opening the file containing the Ideal Chromosome interactions!")
 
-    def _get_chrom_seq(self, filename):
-        R"""Reads the chromatin sequence as letters of the types/compartments.
-        
-        Args:
-
-            filename (file, required):
-                Chromatin sequence of types file. The first column should contain the locus index. The second column should have the locus type annotation. A template of the chromatin sequence of types file can be found at the `Nucleome Data Bank (NDB) <https://ndb.rice.edu/static/text/chr10_beads.txt>`_.
-                
-        Returns:
-            :math:`(N,1)` :class:`numpy.ndarray`:
-                Returns an array of the sequence of chromatin types.
-
-        """
-
-        my_list = []
-        af = open(filename,'r')
-        pos = af.read().splitlines()
-        for t in range(len(pos)):
-            pos[t] = pos[t].split()
-            my_list.append(pos[t][1])
-
-        return np.array(my_list)
-        
+   
     def prob_calculation_IC(self, state):
         R"""
         Calculates the contact probability matrix and the cross term of the Hessian for the Ideal Chromosome optimization.
@@ -667,45 +601,9 @@ class CustomMiChroMTraining:
         """
         return self.PiPj_IC/self.Nframes
 
-    def knight_ruiz_balance(self,matrix, tol=1e-5, max_iter=100):
-        R"""
-        Perform the Knight-Ruiz matrix balancing.
-        """
-        A = np.array(matrix, dtype=float)
-        n = A.shape[0]
-        row_scaling = np.ones(n)
-        col_scaling = np.ones(n)
 
-        for _ in range(max_iter):
-            row_scaling = np.sqrt(np.sum(A, axis=1))
-            A /= row_scaling[:, None]
-            col_scaling = np.sqrt(np.sum(A, axis=0))
-            A /= col_scaling
+      
 
-            if np.all(np.abs(row_scaling - 1) < tol) and np.all(np.abs(col_scaling - 1) < tol):
-                break
-
-        return A
-
-
-    def normalize_matrix(self,matrix):
-        R"""
-        Normalize the matrix for simulation optimization. Here the first neighbor should have the probability of contact P=1.0.
-        """
-        matrix = np.nan_to_num(matrix, nan=0, posinf=0, neginf=0)
-        np.fill_diagonal(matrix,0.0)
-
-        max_values = np.amax(np.triu(matrix), axis=1)
-        
-        # To avoid division by zero, replace zeros with ones
-        max_values[max_values == 0] = 0.0000001
-        
-        normalized_matrix = np.triu(matrix) / max_values[:, np.newaxis]
-        # return normalized_matrix
-        matrix= normalized_matrix + np.triu(normalized_matrix,k=1).T
-        np.fill_diagonal(matrix,1.0)
-
-        return matrix
 
 
     def get_HiC_exp(self, HiC, centerRemove=False, centrange=[0,0], norm=False, cutoff_low=0.0, cutoff_high=1.0, KR=False, neighbors=0):
@@ -732,10 +630,10 @@ class CustomMiChroMTraining:
             allmap = np.loadtxt(HiC)
 
         if KR==True:
-            allmap = knight_ruiz_balance(allmap)
+            allmap = helper.knight_ruiz_balance(allmap)
 
         if norm==True:
-            r=normalize_matrix(allmap)
+            r=helper.normalize_matrix(allmap)
 
             for i in range(len(r)-1):
                 maxElem = r[i][i+1]
