+++ tests/AdamTraining/input/Nframes_0
@@ -1,0 +1,1 @@
+7.500000000000000000e+01
